requirements.txt

streamlit==1.37.1
python-dotenv==1.0.1

langchain==0.2.16
langchain-community==0.2.16
langchain-text-splitters==0.2.2
langchain-groq==0.1.9

faiss-cpu==1.8.0.post1
sentence-transformers==2.7.0
pypdf==4.3.1